Now we have biome transition, and some new models
now biomes have a linear lerp between biomes noises, next is move all code of tiles to another thread, tiles code in main thread is causing some frezzing when generate more complex noise`s, llike the math of biomes transitions

+++ Assets/Scripts/World/MenuChunk.cs
@@ -79,7 +79,7 @@ public class MenuChunk : MonoBehaviour
 
             if (tile.type != TypeBlock.Rock)
             {
-                System.Random randomValue = new System.Random(Game.GameManager.Seed + tile.x + tile.z);
+                System.Random randomValue = new System.Random(Game.GameManager.Seed + (int)tile.x + (int)tile.z);
                 trees.transform.position = new Vector3(tile.x + (float)randomValue.NextDouble(), tile.y, tile.z + (float)randomValue.NextDouble());
             }
